requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.2.4
matplotlib==3.8.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
pytz==2023.3
numpy==1.26.2